requirements.txt
# Core dependencies for simplified news intelligence pipeline

# Data processing
pandas>=1.5.0,<3.0.0
numpy>=1.21.0,<2.0.0
python-dateutil>=2.8.0

# RSS and web scraping
feedparser>=6.0.10
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.6.0

# Google Cloud Platform
google-cloud-storage>=2.5.0,<3.0.0
google-cloud-secret-manager>=2.12.0,<3.0.0
google-auth>=2.16.0,<3.0.0
google-api-python-client>=2.70.0,<3.0.0
google-resumable-media>=2.4.0,<3.0.0

# Email (updated to use requests for Mailgun API)
# sendgrid>=6.9.0  # Removed - replaced with Mailgun
# requests>=2.28.0 already included above for Mailgun API calls

# Machine learning for embeddings
scikit-learn>=1.1.0,<2.0.0

# Social media integrations  
atproto>=0.0.20

# Firestore for configuration
google-cloud-firestore>=2.7.0,<3.0.0

# AI APIs
anthropic>=0.35.0

# Time zone handling
pytz>=2022.1

# Additional utilities
urllib3>=1.26.0,<2.0.0

# Enhanced Sentiment Tracking additions
aiohttp>=3.8.0
asyncio-throttle>=1.0.0
textblob>=0.17.0
nltk>=3.8.0
